+++ lib/quick_start.dart
@@ -16,6 +16,7 @@ export 'infrastructure/network_impl.dart';
 export 'infrastructure/storage_impl.dart';
 
 export 'profile/quick_start_package.dart';
+export 'profile/get_arch_application_x.dart';
 
 export 'package:get_arch_core/get_arch_core.dart';
 export 'package:get_arch_core/profile/i_get_arch_package.dart';